+++ BE/src/main/java/org/team4/airbnb/customer/CustomerController.java
@@ -5,7 +5,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-import org.team4.airbnb.accommodation.dto.AccommodationForWishListResponse;
+import org.team4.airbnb.customer.dto.AccommodationForWishListResponse;
 
 @RestController
 @RequiredArgsConstructor
